Hack to support objects with secondary null value
If an object type is specified as an array and one of those types is "null" then it will now not throw an error.

eg type: ["string", "null"] or type: ["object", "null"]

Note: Values can *already* be empty without throwing an error (unless they are specified as 'required' values) so this changes just increases compatiblity with schemas rather than really changing anything.

This implementation is a bit of a hack - and identifies some existing issues - but works to provide compatiblity with real world schema files like those on opoloproject.com.